re-introduction of route options
after removing the previous route options - which turned out to be
mostly node options - I found that we needed a way to express other
props of the route. I could see this growing, expecially if we want to
add route-level filters (I do). added filters/filter tokens back in and
updated related methods and tests.